modif

+++ rbnb-reservation-service/src/main/java/fr/orleans/univ/miage/m2/rbnbreservationservice/service/ReservationServiceImpl.java
@@ -142,6 +142,9 @@ public class ReservationServiceImpl implements ReservationService {
              ) {
             dateDebutDispo = dispo.getDateDebut();
             dateFinDispo = dispo.getDateFin();
+            if (dateDebutDispo == null || dateFinDispo == null) {
+                throw new NullPointerException();
+            }
 
             if ( ((dateDebutReservation.after(dateDebutDispo))||dateDebutReservation.equals(dateDebutDispo)) && ((dateFinReservation.before(dateFinDispo))||dateFinReservation.equals(dateFinDispo)) ) {
                 disponibiliteRepo.deleteById(dispo.getId());
